get LCs: edge-case workaround
Empirically, ~0.1-1% of lightkurve queries with search_lightcurve raise errors:
SSLError, TimeoutError, ConnectionError, etc.  None of these errors seem to be
easily reproducible.  The implemented workaround here is to just wait a minute,
and try again, for at most three iterations.  If it eventually fails, and
require_lc is false, you'll get an empty list.  Else an assertion error is
raised.